requirements.txt
streamlit~=1.37.1
pydantic~=2.8.2
requests~=2.28.2
pandas~=2.2.2
streamlit_cookies_manager
python-dotenv~=1.0.1
bcrypt~=4.1.3